+++ OzzTradeDiary.NET/OzzTradeDiary.WPF/ViewModels/Maintenance/MaintenanceWindowVM.cs
@@ -1,4 +1,6 @@
-﻿using System.Collections.Specialized;
+﻿using System.Collections.ObjectModel;
+using System.Collections.Specialized;
+using TD.Models;
 using TD.WPF.Commands;
 using TD.WPF.Models;
 
@@ -15,6 +17,8 @@ namespace TD.WPF.ViewModels.Maintenance
 
             _deleteExchangeCommand = new DeleteExchangeCommand(this);
             DeleteExchangeCommand = _deleteExchangeCommand;
+            SymbolExchanges = new ObservableCollection<Exchange>();
+            FilteredSymbols = new ObservableCollection<Symbol>();
 
             PropertyChanged += OnPropertyChanged;
             Symbols.CollectionChanged += OnDependencyCollectionChanged;
@@ -22,6 +26,37 @@ namespace TD.WPF.ViewModels.Maintenance
         }
 
         public DeleteExchangeCommand DeleteExchangeCommand { get; }
+        public ObservableCollection<Symbol> FilteredSymbols { get; }
+        public ObservableCollection<Exchange> SymbolExchanges { get; }
+
+
+        public Exchange? SelectedSymbolExchange
+        {
+            get => _selectedSymbolExchange;
+            set
+            {
+                if (_selectedSymbolExchange?.Id == value?.Id)
+                    return;
+                _selectedSymbolExchange = value;
+                RaisePropertyChanged(nameof(SelectedSymbolExchange));
+                FilterSymbols();
+            }
+        }
+        Exchange? _selectedSymbolExchange = null;
+
+        public string SymbolSearchString
+        {
+            get => _symbolSearchString ?? string.Empty;
+            set
+            {
+                _symbolSearchString = value;
+                RaisePropertyChanged(nameof(SymbolSearchString));
+                FilterSymbols();
+            }
+        }
+        private string? _symbolSearchString;
+
+
 
         public bool CanDeleteSelectedExchange()
         {
@@ -52,6 +87,33 @@ namespace TD.WPF.ViewModels.Maintenance
             _deleteExchangeCommand.RaiseCanExecuteChanged();
         }
 
+        private void FilterSymbols()
+        {
+            if (SelectedSymbolExchange is null && string.IsNullOrWhiteSpace(SymbolSearchString))
+            {
+                ReplaceCollection(FilteredSymbols, Symbols);
+                return;
+            }
+
+            int exchangeId = SelectedSymbolExchange?.Id ?? -1;
+            IEnumerable<Symbol> filtered;
+            if (exchangeId > 0)
+            {
+                filtered = Symbols.Where(symbol => symbol.ExchangeId == exchangeId);
+            }
+            else
+            {
+                filtered = Symbols;
+            }
+
+            if (!string.IsNullOrWhiteSpace(SymbolSearchString))
+            {
+                filtered = filtered.Where(symbol => symbol.TickerFull.Contains(SymbolSearchString, StringComparison.OrdinalIgnoreCase)
+                         || (!string.IsNullOrEmpty(symbol.Description) && symbol.Description.Contains(SymbolSearchString, StringComparison.OrdinalIgnoreCase)));
+            }
+            ReplaceCollection(FilteredSymbols, filtered);
+        }
+
         public async Task LoadAllAsync()
         {
             await LoadCurrenciesAsync();
@@ -59,6 +121,19 @@ namespace TD.WPF.ViewModels.Maintenance
             await LoadTradingAccountsAsync();
             await LoadSymbolsAsync();
             _deleteExchangeCommand.RaiseCanExecuteChanged();
+        }
+
+        public override async Task LoadExchangesAsync()
+        {
+            await base.LoadExchangesAsync();
+            var items = Exchanges.Where(exchange => exchange.HasAnySymbol).ToList();
+            ReplaceCollection(SymbolExchanges, items);
+        }
+
+        public override async Task LoadSymbolsAsync()
+        {
+            await base.LoadSymbolsAsync();
+            FilterSymbols();
         }
 
         private void OnPropertyChanged(object? sender, System.ComponentModel.PropertyChangedEventArgs e)